fix middleware test structure
