+++ tests/Browser/Stubs/BlockListedMiddleware.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Tests\Browser\Stubs;
+
+use Tests\Browser\Security\Component as SecurityComponent;
+
+class BlockListedMiddleware
+{
+    public function handle($request, $next)
+    {
+        SecurityComponent::$loggedMiddleware[] = static::class;
+
+        return $next($request);
+    }
+}
